fix(packaging): Move exceptions.py to tools package to fix installation
