+++ src/tools/exceptions.py
