+++ src/tools/tools.py
@@ -28,7 +28,7 @@ from .tool_params import (
     baseToolArgs,
 )
 from .utils import is_tool_compatible
-from exceptions import HelperOperationError
+from tools.exceptions import HelperOperationError
 from opensearch.helper import (
     add_agentic_memories,
     create_agentic_memory_container,
